feat/7-fetch-and-display-data/fetched data was displayed
