+++ src/components/Recordstable.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { FcNext, FcPrevious } from "react-icons/fc";
-import { db } from "./firebase/firebaseConfig";
+import { db } from "./firebase/firebase";
 import { collection, getDocs, doc, updateDoc } from "firebase/firestore";
 import EditModal from "./EditModal";
