+++ src/pages/AdminPage.jsx
@@ -4,13 +4,14 @@ import SearchBar from "../components/SearchBar";
 import Recordstable from "../components/Recordstable";
 
 export default function AdminPage() {
+  
   return (
     <>
       <div className="backdrop-blur-xl">
         <Header />
         <div className="flex w-screen justify-center gap-20">
           <StyledDatePicker />
-          <SearchBar />
+          <SearchBar  />
         </div>
         <Recordstable />
       </div>
